moved web server to server.py
